LayoutTests/media/modern-media-controls/scrubber-support/scrubber-support-media-api-expected.txt

Testing the ScrubberSupport behavior by using the media API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.timeControl.scrubber.value is 0
PASS mediaController.controls.timeControl.scrubber.value became 0.5

PASS successfullyParsed is true

TEST COMPLETE
